moved intialize.py to games folder so it can be used for all games and created roulette.py
