+++ games/slots_game/slots.py
@@ -1,6 +1,6 @@
 import streamlit as st
 import random
-from .initialize import initial_count
+from ..initialize import initial_count
 
 icons =["🍒", "🍋", "🔔", "💎", "7️⃣"]
